backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
paddlepaddle==2.5.2
paddleocr==2.7.3
PyMuPDF==1.23.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
